Only start handleSend worker after ConnAck
handleSend may attempt to send message from sendCh rather than the
ConnPacket from logicSendC. This is a protocol violation and usually
leads to the broker terminating the connection.

While the user of libmqtt can attempt to avoid this by not sending any
messages before the connection is established, this is not possible in
all cases: When connect() is run as a reconnect, a PublishPacket may
still be buffered in sendCh from before the connection broke.

To avoid this issue, send the ConnPacket bypassing the handleSend logic,
so we can start the handleSend worker later.